+++ packages/plugin-button/src/components/button-sample.jsx
@@ -1,6 +1,6 @@
 import React, { PureComponent } from 'react';
 import PropTypes from 'prop-types';
-import isEqual from 'lodash/isEqual';
+import { isEqual } from 'lodash';
 import { mergeStyles } from 'wix-rich-content-common';
 import PickedIcon from './../icons/pickedIcon';
 import styles from '../../statics/styles/button-sample.scss';