+++ packages/plugin-link/src/LinkViewer.jsx
@@ -1,8 +1,7 @@
 import React, { Component } from 'react';
 import PropTypes from 'prop-types';
 import { normalizeUrl, mergeStyles, validate, Context } from 'wix-rich-content-common';
-import isEqual from 'lodash/isEqual';
-import invoke from 'lodash/invoke';
+import { invoke, isEqual } from 'lodash';
 import schema from '../statics/data-schema.json';
 import styles from '../statics/link-viewer.scss';