+++ examples/main/config/webpack.common.js
@@ -22,16 +22,8 @@ module.exports = env => ({
     extensions: ['.js', '.jsx', '.json'],
     symlinks: false,
     alias: {
-      'react-hot-loader': path.resolve(path.join(__dirname, '..', 'node_modules', 'react-hot-loader')),
-      'draft-js': path.resolve(__dirname, '..', '..', '..', 'node_modules', '@wix', 'draft-js'),
-      '@wix/draft-js': path.resolve(
-        __dirname,
-        '..',
-        '..',
-        '..',
-        'node_modules',
-        '@wix',
-        'draft-js'
+      'react-hot-loader': path.resolve(
+        path.join(__dirname, '..', 'node_modules', 'react-hot-loader')
       ),
       'wix-rich-content-common': path.resolve(__dirname, '..', '..', '..', 'packages', 'common'),
     },