+++ packages/editor/src/RichContentEditor/getBlockRenderMap.js
@@ -2,7 +2,7 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import classNames from 'classnames';
 import { get } from 'lodash';
-import { DefaultDraftBlockRenderMap } from '@wix/draft-js';
+import { DefaultDraftBlockRenderMap } from 'draft-js';
 import { mergeStyles } from 'wix-rich-content-common';
 import styles from '../../statics/styles/rich-content-editor.scss';