+++ packages/plugin-text-color/src/toolbar/TextColorButton.jsx
@@ -2,7 +2,7 @@ import React, { Component } from 'react';
 import PropTypes from 'prop-types';
 import classNames from 'classnames';
 import Modal from 'react-modal';
-import { EditorState } from '@wix/draft-js';
+import { EditorState } from 'draft-js';
 import { InlineToolbarButton, getSelectionStyles, mergeStyles } from 'wix-rich-content-common';
 import TextColorIcon from './TextColorIcon';
 import { TEXT_COLOR_TYPE } from '../types';