+++ packages/plugin-text-color/src/toolbar/TextColorPanel.jsx
@@ -1,6 +1,6 @@
 import React, { Component } from 'react';
 import PropTypes from 'prop-types';
-import { Modifier, EditorState } from '@wix/draft-js';
+import { Modifier, EditorState } from 'draft-js';
 import { ColorPicker, getSelectionStyles } from 'wix-rich-content-common';
 import { DEFAULT_COLOR, DEFAULT_STYLE_SELECTION_PREDICATE } from '../constants';