+++ CHANGELOG.md
@@ -9,6 +9,8 @@
 
 ## [Unreleased]
 ### :bug: Bug Fix
+- `image`
+  - [#342](https://github.com/wix-incubator/rich-content/pull/342) in editor image is not show when helpers are provided
 - `viewer`
   - [139f023](https://github.com/wix-incubator/rich-content/commit/139f023) atomic blocks are no longer rendered in `<p>` tags
 <hr/>