+++ CHANGELOG.md
@@ -8,6 +8,8 @@
 > - :house: [Internal]
 
 ## [Unreleased]
+- `file-upload`
+  - [#345](https://github.com/wix-incubator/rich-content/pull/345) handle empty file name
 <hr/>
 
 ## 4.0.10 (Jul 29, 2019)