+++ CHANGELOG.md
@@ -8,8 +8,11 @@
 > - :house: [Internal]
 
 ## [Unreleased]
+### :bug: Bug Fix
 - `file-upload`
   - [#345](https://github.com/wix-incubator/rich-content/pull/345) handle empty file name
+- `html`
+  - [c2b2101d](https://github.com/wix-incubator/rich-content/commit/c2b2101d) fix height on iphone
 <hr/>
 
 ## 4.0.10 (Jul 29, 2019)