+++ CHANGELOG.md
@@ -8,6 +8,9 @@
 > - :house: [Internal]
 
 ## [Unreleased]
+<hr/>
+
+## 4.0.11 (Jul 31, 2019)
 ### :bug: Bug Fix
 - `file-upload`
   - [#345](https://github.com/wix-incubator/rich-content/pull/345) handle empty file name